feat(mermin-ebpf): adjust integration tests to utilize percpuarray  (#106)
## Description

This is a quick PR to adjust the integration tests for network types to
utilize a PerCpuArray, reflecting the new logic introduced in try_mermin
after we implemented the PerCpuArray with WireGuard. Specifically, I add
a helper function 'store_and_verify_test_data' which writes some data,
including the first byte of parsed data from 'ctx.load', to a
PerCpuArray. It then immediately reads the data to ensure it can be
retrieved.

Each test could be extended to have its own dedicated PerCpuArray
storing that specific header struct, where we write all data to it;
however, I feel that testing the first byte should be sufficient. Please
provide feedback for other specific things we might want to test for the
addition of PerCpuArrays!!