chore(ebpf-refactor): restore integration and unit tests (#126)
## Description
This PR introduces new testing functionality to test tunneled packets,
such as IP-in-IP or VXLAN. The restored tests helped me find one bug in
parse tcp, where we were reading the wrong byte offset for the flag
value. Please double-check this change in mermin-ebpf/main.

Additionally, this PR refactors the network-types integration tests to
more closely align with what we are doing in mermin-ebpf. Specifically,
we remove all the header structs that we no longer use and replace them
with test structs: EthHdr -> EthernetTestData. The TestData structs only
contain the fields from the parsed network header that are actually
extracted in mermin-ebpf.

---------

Co-authored-by: Sven Cowart <svencowart@gmail.com>