fix(ci): Release helm packages to gh-pages (#151)
## Description

While releasig helm archives (packages) to the GH Releases works well
for public repos, for private repos packages should be uploaded to the
`gh-pages` branch itself.

It is related to the complexity of fetching "GH Releases" packages with
the token, which is required for the installing charts from the private
repos.

Also dropping Grafana tests due to failures, chatted IRL /w @reggieross 

## Type of Change

<!-- Please check the box that applies to this pull request. -->

- [x] Bug fix (non-breaking change that resolves an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] Documentation update
- [ ] Refactoring (no functional changes, just code improvements)
- [ ] Other (please describe):

## How Has This Been Tested?

Tested locally without a push

## Checklist

<!-- Confirm the following by checking the boxes. -->

- [x] My code follows the project's coding style guidelines.
- [x] I have performed a self-review of my code.
- [x] I have commented my code, particularly in hard-to-understand
areas.
- [x] I have updated the documentation where necessary.
- [x] My changes generate no new warnings or errors.
- [x] I have added or updated tests that verify my changes.
- [x] All new and existing tests pass.