chore: remove unused tls enabled (#157)
## Description

This PR removes `tls.enabled` configuration option for otlp exporting.
It was unused at the moment but does not align with other configuration
that we implicitly decide when to enable or disable features based on
their relevant configuration flags. In the case of mutual TLS, if the
cert paths are defined then mutual TLS will be enabled.

## Type of Change

<!-- Please check the box that applies to this pull request. -->

- [ ] Bug fix (non-breaking change that resolves an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] Documentation update
- [x] Refactoring (no functional changes, just code improvements)
- [ ] Other (please describe):