fix: Bump Mermin version in the Umbrella chart (#290)
## Changes

* chore: Move "CI" related config to `config.example.hcl`
Purpose of the `config.example.hcl` is for local testing or to be used
as an example for the config for local testing. I don't see strong
reason why we may not set reduced intervals in it, which will not only
benefit local testing, but CI as well
* docs: Remove the redundant doc on Helm
Remove the redundant doc on Helm since volume and mount already present
* docs: Clarify tested platforms
* fix: Bump Mermin version in the Umbrella chart

Fixes #(issue)

### Type of change

- [x] Bug fix
- [ ] New feature
- [ ] Breaking change
- [ ] Documentation
- [ ] Refactor
- [ ] Other:

## Testing

<!-- How did you test this? What environment did you use? -->

## Proof it works

<!-- Screenshots, logs, test output, or other evidence showing your
changes work as expected -->

## Checklist

- [x] I've tested my changes
- [x] I've updated relevant documentation
- [ ] My code follows the project's style (run `cargo fmt` and `cargo
clippy`)
- [x] All tests pass

## Notes

<!-- Anything else reviewers should know? Areas where you'd like
feedback? -->